+++ test/lib/ejsTemplateEngine.js
@@ -6,7 +6,7 @@
 var assert = require("assert");
 
 /**
- * Add general tests for the EjsTemplateEngine to the suite.
+ * Add general tests for the EjsTemplateEngine class to the suite.
  */
 exports.general = function (suite) {
   suite.addBatch({