+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+Work in Progress
+----------------
+
+  * Expanded test coverage to applications with web/socket client tests.
+
 v0.0.8 - 04/06/2013
 -------------------